Fixed a bug in 64bit systems where MPI_Request is a pointer. May produce mismatching communications.

+++ src/merger/paraver/file_set.h
@@ -187,7 +187,7 @@ int SearchRecvEvent_FS (FileSet_t *fset, unsigned int ptask,
 int SearchSendEvent_FS (FileSet_t *fset, unsigned int ptask,
 	unsigned int sender, unsigned int receiver, unsigned int tag,
 	event_t ** send_begin, event_t ** send_end);
-event_t *SearchIRECVED (event_t * current, int request, FileItem_t * freceive);
+event_t *SearchIRECVED (event_t * current, long long request, FileItem_t * freceive);
 
 unsigned int GetTraceOptions (FileSet_t * fset, int numtasks, int taskid);
 int Search_Synchronization_Times (FileSet_t * fset, UINT64 **io_StartingTimes, UINT64 **io_SynchronizationTimes);
